Add migrations for location-aware receiving, items list FIFO index, and kit bundle pricing
- Created Migration_LocationAwareReceiving to implement location-aware receiving with FIFO performance index.
- Created Migration_ItemsListFifoIndex to add performance index for items list FIFO price lookup.
- Created Migration_KitBundlePricing to handle kit bundle pricing and ensure old stock pricing is preserved.
- Added SQL scripts for each migration to define necessary database changes.
- Introduced error handling and language files for stock transfer functionality.
- Developed stock transfer widget and manage views to enhance user experience in transferring stock.
- Implemented tests for FIFO behavior in item batches to ensure pricing integrity across batches.

+++ app/Models/Item_quantity.php
@@ -86,9 +86,14 @@ class Item_quantity extends Model
     /**
      * changes to quantity of an item according to the given amount.
      * if $quantity_change is negative, it will be subtracted,
-     * if it is positive, it will be added to the current quantity
+     * if it is positive, it will be added to the current quantity.
+     *
+     * @param int    $item_id       The item ID.
+     * @param int    $location_id   The stock location ID.
+     * @param float  $quantity_change Amount to add (positive) or subtract (negative).
+     * @return bool True on success.
      */
-    public function change_quantity(int $item_id, int $location_id, int $quantity_change): bool
+    public function change_quantity(int $item_id, int $location_id, float $quantity_change): bool
     {
         $quantity_old = $this->get_item_quantity($item_id, $location_id);
         $quantity_new = $quantity_old->quantity + $quantity_change;
